fixes: #36.
- addcollateral for opencreditaccount wsa not processed as Aftersync hook
in creditmanger passed the txhash as empty string and the multicall
manager had txHash also as empty string as txhash on multicall manager
is set only when there is a multicall with multicallStarted. So, the fix
was to not use txhash from multicall and use LastTxHash for v1 and v2.
- fix the getactiveoraclebyblocknum, also check the version number if
  the blocknum of two oracle adapter is same.